# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  postgres.Comment:
    properties:
      content:
        type: string
      created_at:
        description: Timestamps
        type: string
      edited_at:
        type: string
      id:
        type: integer
      is_edited:
        type: boolean
      like_count:
        type: integer
      likes:
        items:
          $ref: '#/definitions/postgres.CommentLike'
        type: array
      media:
        description: JSON array as string
        type: string
      mentioned_users:
        description: JSON array as string
        type: string
      parent:
        $ref: '#/definitions/postgres.Comment'
      parent_id:
        type: integer
      post:
        allOf:
        - $ref: '#/definitions/postgres.Post'
        description: Relationships
      post_id:
        type: integer
      replies:
        items:
          $ref: '#/definitions/postgres.Comment'
        type: array
      reply_count:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/postgres.User'
      user_id:
        type: integer
    type: object
...
  postgres.Like:
    properties:
      created_at:
        description: Timestamps
        type: string
      id:
        type: integer
      post:
        allOf:
        - $ref: '#/definitions/postgres.Post'
        description: Relationships
      post_id:
        type: integer
      type:
        description: like, love, haha, wow, sad, angry
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/postgres.User'
      user_id:
        type: integer
    type: object
  postgres.Post:
    properties:
      author:
        allOf:
        - $ref: '#/definitions/postgres.User'
        description: Relationships
      author_id:
        type: integer
      comment_count:
        type: integer
      comments:
        items:
          $ref: '#/definitions/postgres.Comment'
        type: array
      content:
        type: string
      created_at:
        description: Timestamps
        type: string
      edited_at:
        type: string
      id:
        type: integer
      is_archived:
        type: boolean
      is_edited:
        type: boolean
      like_count:
        type: integer
      likes:
        items:
          $ref: '#/definitions/postgres.Like'
        type: array
      location:
        type: string
      media:
        items:
          $ref: '#/definitions/postgres.PostMedia'
        type: array
      mentioned_users:
        description: JSON array as string
        type: string
      privacy:
        $ref: '#/definitions/postgres.PostPrivacy'
      share_count:
        type: integer
      shares:
        items:
          $ref: '#/definitions/postgres.Share'
        type: array
      tags:
        description: JSON array as string
        type: string
      type:
        $ref: '#/definitions/postgres.PostType'
      updated_at:
        type: string
      view_count:
        type: integer
      views:
        items:
          $ref: '#/definitions/postgres.PostView'
        type: array
    type: object
...
  postgres.PostPrivacy:
    enum:
    - public
    - friends
    - private
    - public
    - friends
    - private
    - public
    - friends
    - private
    type: string
    x-enum-varnames:
    - PostPrivacyPublic
    - PostPrivacyFriends
    - PostPrivacyPrivate
  postgres.PostType:
    enum:
    - text
    - image
    - video
    - link
    - audio
    - text
    - image
    - video
    - link
    - audio
    - text
    - image
    - video
    - link
    - audio
    type: string
    x-enum-varnames:
    - PostTypeText
    - PostTypeImage
    - PostTypeVideo
    - PostTypeLink
    - PostTypeAudio
...
  postgres.Share:
    properties:
      content:
        description: Optional comment when sharing
        type: string
      created_at:
        description: Timestamps
        type: string
      id:
        type: integer
      post:
        allOf:
        - $ref: '#/definitions/postgres.Post'
        description: Relationships
      post_id:
        type: integer
      privacy:
        $ref: '#/definitions/postgres.PostPrivacy'
      updated_at:
        type: string
      user:
        $ref: '#/definitions/postgres.User'
      user_id:
        type: integer
    type: object
...
  requests.CreateCommentRequest:
    properties:
      content:
        type: string
      mentioned_users:
        items:
          type: integer
        type: array
      parent_id:
        type: integer
    required:
    - content
    type: object
  requests.CreatePostRequest:
    properties:
      content:
        type: string
      location:
        type: string
      media:
        items:
          $ref: '#/definitions/requests.PostMediaRequest'
        type: array
      privacy:
        $ref: '#/definitions/requests.PostPrivacy'
      tags:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/requests.PostType'
    required:
    - type
    type: object
...
  requests.PostMediaRequest:
    properties:
      alt_text:
        type: string
      duration:
        type: integer
      filename:
        type: string
      height:
        type: integer
      mime_type:
        type: string
      order:
        type: integer
      size:
        type: integer
      thumbnail:
        type: string
      type:
        type: string
      url:
        type: string
      width:
        type: integer
    required:
    - type
    - url
    type: object
  requests.PostPrivacy:
    enum:
    - public
    - friends
    - private
    - public
    - friends
    - private
    type: string
    x-enum-varnames:
    - PostPrivacyPublic
    - PostPrivacyFriends
    - PostPrivacyPrivate
  requests.PostType:
    enum:
    - text
    - image
    - video
    - link
    - audio
    - text
    - image
    - video
    - link
    - audio
    type: string
    x-enum-varnames:
    - PostTypeText
    - PostTypeImage
    - PostTypeVideo
    - PostTypeLink
    - PostTypeAudio
...
  requests.SharePostRequest:
    properties:
      content:
        type: string
      privacy:
        $ref: '#/definitions/requests.PostPrivacy'
    type: object
  requests.UpdatePostRequest:
    properties:
      content:
        type: string
      location:
        type: string
      privacy:
        $ref: '#/definitions/requests.PostPrivacy'
      tags:
        items:
          type: string
        type: array
    type: object
...
  responses.CommentResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/postgres.Comment'
        type: array
      next_cursor:
        $ref: '#/definitions/paginator.Cursor'
      total:
        type: integer
    type: object
  responses.PostResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/postgres.Post'
        type: array
      next_cursor:
        $ref: '#/definitions/paginator.Cursor'
      total:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive social media backend with posts, video calling, and
    file uploads
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Social Media Backend API
  version: "2.0"
paths:
...
  /posts:
    post:
      consumes:
      - application/json
      description: Create a new post with text, images, or videos
      parameters:
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/requests.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created post
          schema:
            $ref: '#/definitions/postgres.Post'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - Posts
  /posts/{id}:
    delete:
      description: Delete a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid post ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete post
      tags:
      - Posts
    get:
      description: Retrieve a specific post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Post data
          schema:
            $ref: '#/definitions/postgres.Post'
        "400":
          description: Invalid post ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
      summary: Get post by ID
      tags:
      - Posts
    put:
      consumes:
      - application/json
      description: Update an existing post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Post update data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/requests.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated post
          schema:
            $ref: '#/definitions/postgres.Post'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update post
      tags:
      - Posts
  /posts/{id}/comments:
    get:
      description: Get comments for a specific post with cursor-based pagination
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - default: 20
        description: Number of comments to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comments data
          schema:
            $ref: '#/definitions/responses.CommentResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
      summary: Get post comments
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: Create a comment on a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment data
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/requests.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created comment
          schema:
            $ref: '#/definitions/postgres.Comment'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create comment
      tags:
      - Posts
  /posts/{id}/like:
    post:
      description: Toggle like status on a post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - default: '"like"'
        description: Type of like
        in: query
        name: like_type
        type: string
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Like/Unlike post
      tags:
      - Posts
  /posts/{id}/share:
    post:
      consumes:
      - application/json
      description: Share a post to user's timeline
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Share data
        in: body
        name: share
        required: true
        schema:
          $ref: '#/definitions/requests.SharePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created share
          schema:
            $ref: '#/definitions/postgres.Share'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Share post
      tags:
      - Posts
  /posts/feed:
    get:
      description: Get personalized feed with posts from friends
      parameters:
      - default: 20
        description: Number of posts to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Feed data
          schema:
            $ref: '#/definitions/responses.PostResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user feed
      tags:
      - Posts
  /posts/public:
    get:
      description: Get public posts for explore page
      parameters:
      - default: 20
        description: Number of posts to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Public feed data
          schema:
            $ref: '#/definitions/responses.PostResponse'
      summary: Get public feed
      tags:
      - Posts
  /posts/search:
    get:
      description: Search posts by content or tags
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Number of posts to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search results
          schema:
            $ref: '#/definitions/responses.PostResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      summary: Search posts
      tags:
      - Posts
  /posts/user/{user_id}:
    get:
      description: Get posts by a specific user with cursor-based pagination
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - default: 20
        description: Number of posts to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Posts data
          schema:
            $ref: '#/definitions/responses.PostResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
      summary: Get user posts
      tags:
      - Posts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...